modification deplacement de la fonction menu

+++ getSambaScript.sh
@@ -1,5 +1,15 @@
 #!/bin/bash
 
+function menu {
+    echo "Que faire ?"
+    echo "1. Désinstaller Samba"
+    echo "2. Ajouter un utilisateur"
+    echo "3. Ajouter un disque partagé"
+    echo "4. Voir les dossiers partagés"
+    echo "5. Voir les utilisateurs et les mots de passe"
+    echo "6. Quitter"
+}
+
 # Vérifier si Samba est installé
 if ! command -v smbd &> /dev/null; then
     echo "Samba n'est pas installé. Installation en cours..."
@@ -14,22 +24,17 @@ if ! command -v smbd &> /dev/null; then
     sudo /etc/init.d/smbd restart
     echo "Samba installé, les dossiers partagés sont dans /home/shares/public."
     menu
-    ##si samba est deja installer alors on affiche le menu
+    ## Si Samba est déjà installé, afficher le menu
 else
     menu
 fi
 
-function menu {
-    echo "que faire ?"
-    echo "1. Désinstaller Samba"
-    echo "2. Ajouter un utilisateur"
-    echo "3. Ajouter un disque partagé"
-    echo "4. voir les dossiers partagés"
-    echo "5. voir les utilisateurs et les mots de passe"
-    echo "6. Quitter"
-}
-
 read -p "Votre choix : " choix
+
+while [[ ! $choix =~ ^[1-6]$ ]]; do
+    echo "Choix invalide. Veuillez entrer un nombre entre 1 et 6."
+    read -p "Votre choix : " choix
+done
 
 if [ $choix -eq 1 ]; then
     echo "Désinstallation de Samba..."
@@ -49,31 +54,31 @@ elif [ $choix -eq 3 ]; then
     echo "Ajout d'un disque partagé..."
     read -p "Nom du disque (exemple disk1=sda1 / disk2 =sda2): " disk
     sudo mkdir /home/shares/public/$disk
-    #verifier si le disque a un chiffre dans son nom si oui alors on separer le nom du disque et le chiffre pour pouvoir le mettre a sda
-    if [[ $disk =~ [0-9] ]]; then
-        diskName=$(echo $disk | cut -d [0-9] -f1)
-        diskNumber=$(echo $disk | cut -d [0-9] -f2)
-        sda="sda$diskNumber"
-    fi
-    sudo mount /dev/$sda /home/shares/public/$disk
 
-    sudo cp fstab /etc/fstab
+    # Vérifier si le disque existe
+    if [[ -e /dev/$disk ]]; then
+        # Vérifier si le disque a un chiffre dans son nom
+        if [[ $disk =~ [0-9] ]]; then
+            diskName=$(echo $disk | cut -d [0-9] -f1)
+            diskNumber=$(echo $disk | cut -d [0-9] -f2)
+            sda="sda$diskNumber"
+        fi
+        sudo mount /dev/$sda /home/shares/public/$disk
+        sudo cp fstab /etc/fstab
+        echo "Disque partagé ajouté."
+    else
+        echo "Le disque n'existe pas."
+    fi
     exit 0
 elif [ $choix -eq 4 ]; then
     echo "Liste des dossiers partagés :"
     sudo cat /etc/samba/smb.conf | grep "\[" | cut -d "[" -f2 | cut -d "]" -f1
     exit 0
-
 elif [ $choix -eq 5 ]; then
     echo "Liste des utilisateurs et des mots de passe :"
     sudo cat /etc/samba/smbpasswd
     exit 0
 else
-    echo "Choix invalide."
+    echo "Script terminé."
     exit 1
 fi
-
-# Démarrer le service Samba
-sudo service smbd start
-
-echo "Script terminé."
